Falta el perfil

+++ grails-app/domain/ar/unlam/edu/Lista.groovy
@@ -1,7 +1,7 @@
 package ar.unlam.edu
 
 class Lista {
-	Long listaOriginalId
+	Integer listaOriginalId
 	String name
 	SecUser usuario
 	static hasMany = [canciones:Cancion]
